Adding the new Ranged Enemy Functionality
I have the zombie disabled right now because I'm modularizing the AI Controller scripts and working on the Ranged Enemy.
