+++ Assets/Animation/Enemy/Mailbot.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 295863006d8486548b2773fe7a92b476
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
