+++ Assets/Animation/Enemy/Mailbot/Idle.anim.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: c0f20d4c1ba0d1c43bf7c64b120137b5
+NativeFormatImporter:
+  externalObjects: {}
+  mainObjectFileID: 7400000
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
