+++ Assets/Animation/Enemy/Mailbot/Mailbot.controller.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 59dae7be41b5c5947a16706a8f490aab
+NativeFormatImporter:
+  externalObjects: {}
+  mainObjectFileID: 9100000
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
