+++ Assets/Animation/Enemy/Mailbot/Throwing.anim.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: b6db4dd2e127c3c4b949b64711922522
+NativeFormatImporter:
+  externalObjects: {}
+  mainObjectFileID: 7400000
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
