+++ Assets/Animation/Enemy/Mailbot/Walking.anim.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 973f2434a1b052e42950ae4416d81603
+NativeFormatImporter:
+  externalObjects: {}
+  mainObjectFileID: 7400000
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
