+++ Assets/Animation/Enemy/Zombie.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 457cbcb82a28da14a903e9952f212e97
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
