+++ Assets/Animation/Enemy/Zombie/Idle.anim
