+++ Assets/Prefabs/Enemy/Prototypes/Mailbot.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: dbfb2eb53828a0a4f86bedd6b18cef2f
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
