+++ Assets/Prefabs/Enemy/Prototypes/Mailbot/Mailbot.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: c24e4ef189235364a99fa813d1a849c9
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
