+++ Assets/Prefabs/Enemy/Prototypes/Mailbot/Mailbot/Mailbot.mat.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 38dc17c0cdba5ef4ab2226f7edbafbaa
+NativeFormatImporter:
+  externalObjects: {}
+  mainObjectFileID: 2100000
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
