+++ Assets/Prefabs/Enemy/Prototypes/Zombie.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 66613f2f5796b6b4b886322997667e38
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
