+++ Assets/Prefabs/Enemy/Prototypes/Zombie/GirlScoutZombie_Material.mat
