+++ Assets/Prefabs/Enemy/Prototypes/Zombie/GirlScoutZombie_Material1.mat.meta
