+++ Assets/Prefabs/Enemy/Prototypes/Zombie/GirlScout_eye.png.meta
