+++ Assets/Scenes/GavinsScene/NavMesh-Terrain.asset.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 567d8bc04f5dfc04e867d93ddee97518
+NativeFormatImporter:
+  externalObjects: {}
+  mainObjectFileID: 23800000
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
