+++ Assets/Scripts/Enemy/ShortRangeEnemy.cs.meta
@@ -1,0 +1,2 @@
+fileFormatVersion: 2
+guid: d1d9a120d7a59a24798083d8ee0cf1a1
